Massive ddev refactor
