+++ datadog_checks_dev/datadog_checks/dev/__init__.py
@@ -2,10 +2,5 @@
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
 from .__about__ import __version__
-from .conditions import WaitFor
-from .docker import docker_run, get_docker_hostname
-from .env import environment_run
 from .errors import RetryError
-from .structures import EnvVars, LazyFunction, TempDir
 from .subprocess import run_command
-from .utils import chdir, get_here, load_jmx_config, temp_chdir, temp_dir
