+++ datadog_checks_dev/datadog_checks/dev/testing/constants.py
@@ -6,3 +6,4 @@
 # a test will run against the check containing the file
 TESTABLE_FILE_PATTERNS = ('*.py', '*.ini', '*.in', '*.txt', '*.yml', '*.yaml', '**/tests/*')
 NON_TESTABLE_FILES = ('auto_conf.yaml', 'agent_requirements.in')
+REQUIREMENTS_IN = 'requirements.in'
