+++ datadog_checks_dev/datadog_checks/dev/testing/e2e/README.md
@@ -1,0 +1,1 @@
+This package contains utils for ddev[cli] e2e testing tools
