+++ datadog_checks_dev/datadog_checks/dev/testing/environment/docker.py
@@ -5,16 +5,19 @@ import os
 from contextlib import contextmanager
 from typing import Iterator
 
+from datadog_checks.dev import run_command
+from datadog_checks.dev.fileutils import file_exists, create_file
+from datadog_checks.dev.spec import load_spec
+from datadog_checks.dev.testing.conditions import CheckDockerLogs
+from datadog_checks.dev.testing.environment.run import environment_run
+from datadog_checks.dev.testing.environment.structures import TempDir, EnvVars, LazyFunction
+from datadog_checks.dev.testing.environment.utils import get_state, save_state
+from datadog_checks.dev.testing.utils import find_check_root
 from six import string_types
 from six.moves.urllib.parse import urlparse
 from tenacity import retry, stop_after_attempt, wait_fixed
 
-from .conditions import CheckDockerLogs
-from .env import environment_run, get_state, save_state
-from .spec import load_spec
-from .structures import EnvVars, LazyFunction, TempDir
-from .subprocess import run_command
-from .utils import create_file, file_exists, find_check_root
+
 
 try:
     from contextlib import ExitStack
