+++ datadog_checks_dev/datadog_checks/dev/tooling/constants.py
@@ -84,8 +84,6 @@ NOT_TILES = [
 ]
 
 
-REQUIREMENTS_IN = 'requirements.in'
-
 ROOT = ''
